requirements.txt
#main modules
django >3.2,<3.3
python-decouple
Pillow
djangorestframework

#third party modules
markdown       
django-filter  
drf-yasg[validation]
djangorestframework-simplejwt


#email third party modules
django-mail-templated


#reformatiing and tests
black
flake8